// src/core/models/user/userDetail.model.ts
export class UserDetail {
    avatar: string;
    full_name: string;
    position: string;
    phone: string;
    organ: string;
    group: string;
    birthDay: string;
    number: string;
    nation: string;
    sex: string;
    hometown: string;
    religion: string;
    preparatoryDay: string;
    officialDay: string;
    Education: string;
    degree: string;
    note: string;
    specialize: string;
    discipline: string;
    Bonus: string;
    politicalTheory: string;
}